+++ src/app/(shop)/products/_components/ProductsFilters.tsx
@@ -1,248 +1,16 @@
-'use client'
-
-import { useSearchParams, useRouter, usePathname } from 'next/navigation'
-import { useOptimistic, useTransition } from 'react'
-import useCreateQueryString from '@/hooks/use-create-query-string'
-import Link from 'next/link'
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from '@/components/ui/dropdown-menu'
-import { Button } from '@/components/ui/button'
-import { PaginationDisplayValues } from '@/lib/types'
 import { Separator } from '@/components/ui/separator'
-import { Switch } from '@/components/ui/switch'
+import TrendingFilter from '@/components/custom/filters/TrendingFilter'
+import DisplayFilter from '@/components/custom/filters/DisplayFilter'
+import SortFilter from '@/components/custom/filters/SortFilter'
 
 export default function ProductsFilters() {
-  const router = useRouter()
-  const searchParams = useSearchParams()
-  const [isPending, startTransition] = useTransition()
-  const pathname = usePathname()
-  const [optimisticSort, setOptimisticSort] = useOptimistic(
-    searchParams.get('sortiranje')
-  )
-  const [optimisticDisplay, setOptimisticDisplay] = useOptimistic(
-    searchParams.get('prikazi')
-  )
-  const [optimisticTrending, setOptimisticTrending] = useOptimistic(
-    searchParams.get('aktuelno')
-  )
-
-  const createQueryString = useCreateQueryString(searchParams)
-
-  const paginationDisplayOptions = [
-    {
-      label: PaginationDisplayValues.small,
-      url: {
-        pathname,
-        query: createQueryString({
-          addParams: [
-            { name: 'prikazi', value: PaginationDisplayValues.small },
-          ],
-          removeParams: ['stranica'],
-        }),
-      },
-      value: PaginationDisplayValues.small,
-    },
-    {
-      label: PaginationDisplayValues.medium,
-      url: {
-        pathname,
-        query: createQueryString({
-          addParams: [
-            { name: 'prikazi', value: PaginationDisplayValues.medium },
-          ],
-          removeParams: ['stranica'],
-        }),
-      },
-      value: PaginationDisplayValues.medium,
-    },
-    {
-      label: PaginationDisplayValues.large,
-      url: {
-        pathname,
-        query: createQueryString({
-          addParams: [
-            { name: 'prikazi', value: PaginationDisplayValues.large },
-          ],
-          removeParams: ['stranica'],
-        }),
-      },
-      value: PaginationDisplayValues.large,
-    },
-    {
-      label: PaginationDisplayValues.extraLarge,
-      url: {
-        pathname,
-        query: createQueryString({
-          addParams: [
-            { name: 'prikazi', value: PaginationDisplayValues.extraLarge },
-          ],
-          removeParams: ['stranica'],
-        }),
-      },
-      value: PaginationDisplayValues.extraLarge,
-    },
-  ]
-
-  const sortingOptions = [
-    {
-      label: 'Najnovijim',
-      url: {
-        pathname,
-        query: createQueryString({
-          addParams: [{ name: 'sortiranje', value: 'najnoviji' }],
-        }),
-      },
-      value: 'najnoviji',
-    },
-    {
-      label: 'Najnižoj ceni',
-      url: {
-        pathname,
-        query: createQueryString({
-          addParams: [{ name: 'sortiranje', value: 'najniza-cena' }],
-        }),
-      },
-      value: 'najniza-cena',
-    },
-    {
-      label: 'Najvišoj ceni',
-      url: {
-        pathname,
-        query: createQueryString({
-          addParams: [{ name: 'sortiranje', value: 'najvisa-cena' }],
-        }),
-      },
-      value: 'najvisa-cena',
-    },
-  ]
-
-  let orderingLabel: string
-
-  switch (optimisticSort) {
-    case 'najvisa-cena':
-      orderingLabel = 'Najvišoj ceni'
-      break
-    case 'najniza-cena':
-      orderingLabel = 'Najnižoj ceni'
-      break
-    case 'najnoviji':
-      orderingLabel = 'Najnovijim'
-      break
-    default:
-      orderingLabel = 'Najnovijim'
-  }
-
-  let displayLabel: string
-
-  switch (optimisticDisplay) {
-    case PaginationDisplayValues.small:
-      displayLabel = PaginationDisplayValues.small
-      break
-    case PaginationDisplayValues.medium:
-      displayLabel = PaginationDisplayValues.medium
-      break
-    case PaginationDisplayValues.large:
-      displayLabel = PaginationDisplayValues.large
-      break
-    case PaginationDisplayValues.extraLarge:
-      displayLabel = PaginationDisplayValues.extraLarge
-      break
-    default:
-      displayLabel = PaginationDisplayValues.small
-  }
-
-  let trending
-
-  switch (optimisticTrending) {
-    case 'da':
-      trending = true
-      break
-    case 'ne':
-      trending = false
-      break
-    default:
-      trending = false
-  }
-
-  const handleTrendingChange = (checked: boolean) => {
-    startTransition(() => {
-      setOptimisticTrending(checked ? 'da' : 'ne')
-    })
-    router.push(
-      `${pathname}?${createQueryString({
-        addParams: [{ name: 'aktuelno', value: checked ? 'da' : 'ne' }],
-        removeParams: ['stranica'],
-      })}`
-    )
-  }
-
   return (
-    <div
-      data-pending-products={isPending ? '' : undefined}
-      className='space-y-4 md:space-y-0 md:flex md:space-x-4 items-center'
-    >
-      <div className='flex items-center'>
-        <span className='mr-2'>Samo aktuelno:</span>
-        <Switch
-          checked={trending}
-          onCheckedChange={handleTrendingChange}
-          aria-label='trending-switch'
-        />
-      </div>
+    <div className='space-y-4 md:space-y-0 md:flex md:space-x-4 items-center'>
+      <TrendingFilter />
       <Separator className='md:hidden' />
-      <div>
-        <span className='mr-2'>Prikazi:</span>
-        <DropdownMenu>
-          <DropdownMenuTrigger asChild>
-            <Button variant='secondary'>{displayLabel}</Button>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent>
-            {paginationDisplayOptions.map((option) => (
-              <DropdownMenuItem key={option.label} asChild>
-                <Link
-                  href={option.url}
-                  onClick={() => {
-                    startTransition(() => {
-                      setOptimisticDisplay(option.value)
-                    })
-                  }}
-                >
-                  {option.label}
-                </Link>
-              </DropdownMenuItem>
-            ))}
-          </DropdownMenuContent>
-        </DropdownMenu>
-      </div>
+      <DisplayFilter />
       <Separator className='md:hidden' />
-      <div>
-        <span className='mr-2'>Sortiranje prema:</span>
-        <DropdownMenu>
-          <DropdownMenuTrigger asChild>
-            <Button variant='secondary'>{orderingLabel}</Button>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent>
-            {sortingOptions.map((option) => (
-              <DropdownMenuItem key={option.label} asChild>
-                <Link
-                  href={option.url}
-                  onClick={() => {
-                    startTransition(() => {
-                      setOptimisticSort(option.value)
-                    })
-                  }}
-                >
-                  {option.label}
-                </Link>
-              </DropdownMenuItem>
-            ))}
-          </DropdownMenuContent>
-        </DropdownMenu>
-      </div>
+      <SortFilter />
     </div>
   )
 }
